pyproject.toml
[project]
name = "test-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.3.0",
    "faiss",
    "ipykernel>=7.1.0",
    "langchain>=1.0.2",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.1",
    "langchain-groq>=1.0.0",
    "langchain-text-splitters>=1.0.0",
    "networkx>=3.5",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pymupdf>=1.26.5",
    "pypdf>=6.1.3",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.1.2",
    "unstructured>=0.18.15",
]
